Renamed and added new models
After some research I learned that what I used to call AlexNets actually
were no AlexNets but something else (confusion originated in what turned
out to be an unreliable source). I renamed those models mynet and created
a new model called AlexNet which uses the actualy AlexNet architecture
with the single exception that it uses just one instead of two dense layers
at the end call AlexNets actually
were no AlexNets but something else (confusion originated in what turned
out to be an unreliable source). I renamed those models mynet and created
a new model called AlexNet which uses the actualy AlexNet architecture
with the single exception that it uses just one instead of two dense layers
at the end. This modification was introduced in order to significantly
reduce the number of parameters of the model which was necessary to train
the model on my MacBook Pro with only 8GB RAM.